Webpack refactoring: Share common parts.
