+++ scenarioo-client/configs/webpack.config.prod.js
@@ -1,0 +1,9 @@
+var merge = require('webpack-merge');
+var commonConfig = require('./webpack.config.common');
+
+var webpackConfig = merge(commonConfig, {
+
+});
+
+
+module.exports = webpackConfig;
